server/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
pydantic==2.7.4
python-dotenv==1.0.1
aiosmtplib==2.0.2
email-validator==2.2.0
